Reorganized code into disparate files.
