+++ lib/acts_as_paranoid/validations.rb
@@ -1,7 +1,12 @@
 require 'active_support/core_ext/array/wrap'
 
-module ParanoidValidations
-  class UniquenessWithoutDeletedValidator < ActiveRecord::Validations::UniquenessValidator
+module ActsAsParanoid
+  module Validations
+    def self.included(base)
+      base.extend ClassMethods
+    end
+
+    class UniquenessWithoutDeletedValidator < ActiveRecord::Validations::UniquenessValidator
       def validate_each(record, attribute, value)
         finder_class = find_finder_class_for(record)
         table = finder_class.arel_table
@@ -26,11 +31,12 @@ module ParanoidValidations
           record.errors.add(attribute, :taken, options.except(:case_sensitive, :scope).merge(:value => value))
         end
       end
-  end
+    end
 
-  module ClassMethods
-    def validates_uniqueness_of_without_deleted(*attr_names)
-      validates_with UniquenessWithoutDeletedValidator, _merge_attributes(attr_names)
+    module ClassMethods
+      def validates_uniqueness_of_without_deleted(*attr_names)
+        validates_with UniquenessWithoutDeletedValidator, _merge_attributes(attr_names)
+      end
     end
   end
 end
